pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "pybind11","scikit-build-core"]
build-backend = "scikit_build_core.build"

[tool.pytest.ini_options]
testpaths = ["test/python"]
addopts = "-v --disable-warnings"

[tool.uv.sources]
manipulation = { git = "https://github.com/RussTedrake/manipulation.git" }

[project]
name = "drake-final-project"
version = "0.1.0"
description = "Drake environment for final project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "meshcat>=0.3.2",
    "numpy>=2.2.6",
    "pybind11>=3.0.1",
    "scipy>=1.15.3",
    "setuptools>=80.9.0",
    "scikit-build-core",
    "pydrake>=0.2.1",
    "drake>=1.46.0",
    "hydra-core>=1.3.2",
    "manipulation",
]
